Separate symbol class tests from tests for symbol definitions
